backend/requirements.txt
asgiref==3.6.0
Django==3.2.18
django-filter==21.1
djangorestframework==3.14.0
drf-base64==2.0
fpdf==1.7.2
gunicorn==20.1.0
isort==5.11.4
Pillow==9.4.0
psycopg2-binary==2.9.5
pytz==2022.7.1
reportlab==3.6.12
sqlparse==0.4.3
python-dotenv==0.20.0
djoser==2.1.0